Make global refactoring for search controller
Make SearchController as MainCollectionViewController's instance
Simplify SearchResultsCollectionViewController
