+++ iTunesApp/Application/Controllers/Main/MainCollectionViewController+DataSource.swift
@@ -8,6 +8,7 @@
 import UIKit
 
 extension MainCollectionViewController {
+  
   func configureDataSource() {
     dataSource = DataSource(collectionView: collectionView) {
       (collectionView, indexPath, item) -> UICollectionViewCell? in
