+++ iTunesApp/Application/Controllers/Main/MainCollectionViewController.swift
@@ -7,55 +7,89 @@
 
 import UIKit
 
-
-
-
-
-/* The Vision of Hell `app` */
-// TODO: create initializer
 final class MainCollectionViewController: UICollectionViewController {
 
-  // MARK: - Typealiases
+  // MARK: - Type Definitions
+  
+  typealias DataSource = UICollectionViewDiffableDataSource<Section, StoreItem>
+  typealias Snapshot = NSDiffableDataSourceSnapshot<Section, StoreItem>
 
-  typealias DataSource = UICollectionViewDiffableDataSource<String, StoreItem>
-  typealias Snapshot = NSDiffableDataSourceSnapshot<String, StoreItem>
+  // MARK: - Sections
 
-  // MARK: - Properties
+  enum Section {
+    case storeItems
+  }
 
-  var items = [StoreItem]()
+  // MARK: - Collection State Manager
+
+  private enum MainCollectionStateManager {
+    case loading
+    case loaded(items: [StoreItem])
+    case info(message: String)
+  }
+
+  // MARK: - Collection View Properties
+
+  private let coreDataManager = DataManager()
+
+  private var items = [StoreItem]()
+
   var dataSource: DataSource!
-
-  // MARK: - Class Instances
-
-  private let dataManager = DataManager()
-  var searchController = SearchController()
-  var searchRepository: SearchRepository
-  lazy var activityIndicator = ActivityIndicator(view: view, style: .large)
-  private lazy var errorView = ErrorView(frame: view.bounds)
-
-  // MARK: - Snapshot
-
-  var itemSnapshot: Snapshot {
+  
+  private var itemSnapshot: Snapshot {
     var snapshot = Snapshot()
-    snapshot.appendSections(["Results"])
+    snapshot.appendSections([.storeItems])
     snapshot.appendItems(items)
     return snapshot
   }
 
-  // FIXME: - add AsyncSequence
-  var imageLoadTask: [IndexPath: Task<Void, Never>] = [:]
-
-  //MARK: - Initializer
-
-  override init(collectionViewLayout: UICollectionViewLayout) {
-    searchRepository = searchController.searchRepository
-    super.init(collectionViewLayout: collectionViewLayout)
+  private var collectionState: MainCollectionStateManager = .info(message: "Введите запрос") {
+    didSet {
+      updateUI()
+    }
   }
+
+  // MARK: - Private Properties
+
+  // MARK: - Additional Views
+  private lazy var activityIndicator = ActivityIndicator(view: view, style: .large)
+  private lazy var errorView = ErrorView(frame: view.bounds)
   
+
+  // MARK: - Network Properties
+  var searchRepository = SearchRepository()
+  var workItem: DispatchWorkItem?
+
+
+  // MARK: - SearchController
+
+  var searchController: UISearchController
+  var resultsController: SearchResultsCollectionViewController
+  var queryOptions = SearchScope.allCases.map{ $0.mediaType }
+
+  var textDidChange: ((String) -> Void)?
+
+  lazy var responseData = ResponseData(
+    service: .storeItems,
+    mediaType: queryOptions[searchController.searchBar.selectedScopeButtonIndex],
+    elementsAmount: 10
+  )
+
+  // MARK: - Initializers
+
+  init() {
+    resultsController = SearchResultsCollectionViewController(
+      searchRepository,
+      coreDataManager: coreDataManager
+    )
+    searchController = UISearchController(searchResultsController: resultsController)
+    super.init(collectionViewLayout: UICollectionViewLayout())
+  }
+
   required init?(coder: NSCoder) {
     fatalError("init(coder:) has not been implemented")
   }
-  
+
   // MARK: - viewDidLoad
 
   override func viewDidLoad() {
@@ -66,105 +100,66 @@ final class MainCollectionViewController: UICollectionViewController {
       forCellWithReuseIdentifier: MainCollectionViewCell.ID
     )
 
+    setupSearchBar()
+    setupResultsController()
+    prepareUI()
     configureColectionView()
     configureDataSource()
-    prepareUI()
   }
 
-  override func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-    let vc = DetailViewController(itemDetails: items[indexPath.item])
-    dataManager.save(item: items[indexPath.item])
-    navigationController?.pushViewController(vc, animated: true)
-  }
-
-  // MARK: - Prepare UI
-
-  func prepareUI() {
+  private func prepareUI() {
     collectionView.backgroundColor = .systemBackground
     view.addSubview(errorView)
   }
 
-  // MARK: - Configure CollectionView
-
-  func configureColectionView() {
+  private func configureColectionView() {
     navigationItem.searchController = searchController
-//    resultSearchController.searchBar.delegate = self
     navigationItem.hidesSearchBarWhenScrolling = false
     collectionView.collectionViewLayout = configureLayout()
-    searchController.searchRepository.delegate = self
+    searchRepository.delegate = self
   }
 
-  // MARK: - Layout
-
-  func configureLayout() -> UICollectionViewLayout {
-    let sectionSpacing: CGFloat = 16
-    let innerSpacing: CGFloat = 14
-
-    let itemSize = NSCollectionLayoutSize(
-      widthDimension: .fractionalWidth(0.5),
-      heightDimension: .fractionalWidth(0.5)
-    )
-    let item = NSCollectionLayoutItem(layoutSize: itemSize)
-
-    let groupSize = NSCollectionLayoutSize(
-      widthDimension: .fractionalWidth(1),
-      heightDimension: .fractionalWidth(0.5)
-    )
-    let group = NSCollectionLayoutGroup.horizontal(
-      layoutSize: groupSize,
-      repeatingSubitem: item,
-      count: 2
-    )
-    group.interItemSpacing = NSCollectionLayoutSpacing.fixed(innerSpacing)
-
-    let section = NSCollectionLayoutSection(group: group)
-    section.contentInsets = NSDirectionalEdgeInsets(
-      top: sectionSpacing,
-      leading: sectionSpacing,
-      bottom: sectionSpacing,
-      trailing: sectionSpacing
-    )
-    section.interGroupSpacing = innerSpacing
-
-    return UICollectionViewCompositionalLayout(section: section)
+  private func updateUI() {
+    switch collectionState {
+    case .loading:
+      errorView.isHidden = true
+      activityIndicator.showIndicator()
+      return
+    case .loaded(let items):
+      errorView.isHidden = true
+      self.items = items
+    case .info(let message):
+      errorView.isHidden = false
+      let errorContext = ErrorContext(message: message)
+      errorView.configureController(errorContext)
+    }
+    activityIndicator.hideIndicator()
   }
 
-  // MARK: - Screen Management
+  // MARK: - Override Methods
 
-  private func showCollectionView(with items: [StoreItem]) {
-    errorView.isHidden = true
-    self.items = items
+  override func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+    let vc = DetailViewController(itemDetails: items[indexPath.item])
+    coreDataManager.save(item: items[indexPath.item])
+    navigationController?.pushViewController(vc, animated: true)
   }
 
-  private func showLoadingView() {
-    errorView.isHidden = true
-    activityIndicator.showIndicator()
-  }
-
-  private func showErrorView(with message: String) {
-    errorView.isHidden = false
-
-    let errorContext = ErrorContext(message: message)
-    errorView.configureController(errorContext)
-  }
+  // MARK: - Private Methods
 }
 
 extension MainCollectionViewController: SearchRepositoryDelegate {
 
   func update() {
     self.items = []
-    switch searchController.searchRepository.storeStateManager.state {
+    switch searchRepository.storeStateManager.state {
     case .empty:
-      activityIndicator.hideIndicator()
-      showErrorView(with: "Введи запрос")
+      collectionState = .info(message: "Введите запрос")
     case .loading:
-      showLoadingView()
+      collectionState = .loading
     case .loaded(let items):
-      self.activityIndicator.hideIndicator()
-      self.showCollectionView(with: items)
+      collectionState = .loaded(items: items)
     case .error(let error):
-      activityIndicator.hideIndicator()
-      showErrorView(with: error.customDescription)
+      collectionState = .info(message: error.customDescription)
     }
     dataSource.apply(itemSnapshot, animatingDifferences: true)
   }
