+++ iTunesApp/Application/Controllers/SearchResults/SectionHeaderView.swift
